// Case_Study/case-study/src/app/model/employee.ts
export class Employee {
  id: number;
  idEmployee: string;
  name: string;
  dayOfBirth: string;
  email: string;
  address: string;
  idCard: string;
  phone: number;
  salary: number;

  constructor() {
    this.id = 0;
    this.idEmployee = '';
    this.name = '';
    this.dayOfBirth = '';
    this.email = '';
    this.address = '';
    this.idCard = '';
    this.phone = 0;
    this.salary = 0;
  }
}